moved gauge reward total to hook
